Style fixes, cleanups, file renaming
