+++ query/pipeline/interfaces/PipelineEdgeOutputInterface.h
@@ -33,10 +33,6 @@ public:
                                                         _edgeTypes->getTag());
     }
 
-    void setPathTag(ColumnTag tag) {
-        _stream.setEdgePathTag(tag);
-    }
-
     void setIndices(NamedColumn* indices) { _indices = indices; }
 
     NamedColumn* getIndices() const { return _indices; }
