+++ query/pipeline/processors/MaterializeData.cpp
@@ -7,8 +7,6 @@
 
 #include "PipelineException.h"
 
-#include "GraphPath.h"
-
 using namespace db;
 
 MaterializeData::MaterializeData(LocalMemory* mem, DataframeManager* dfMan)
@@ -106,7 +104,6 @@ INSTANTIATE(ColumnVector<types::String::Primitive>);
 INSTANTIATE(ColumnVector<types::Bool::Primitive>);
 INSTANTIATE(ColumnVector<types::Embedding::Primitive>);
 
-INSTANTIATE(ColumnVector<Path>);
 INSTANTIATE(ColumnVector<EntityList>);
 
 INSTANTIATE(ColumnOptVector<types::UInt64::Primitive>);
