+++ query/pipeline/processors/MaterializeProcessor.cpp
@@ -14,8 +14,6 @@
 
 #include "BioAssert.h"
 
-#include "GraphPath.h"
-
 using namespace db;
 
 namespace {
@@ -48,7 +46,6 @@ inline void copyChunkImpl(const Column* srcPtr,
         COPY_CHUNK_CASE(ColumnOptVector<types::Double::Primitive>)
         COPY_CHUNK_CASE(ColumnOptVector<types::String::Primitive>)
         COPY_CHUNK_CASE(ColumnOptVector<types::Bool::Primitive>)
-        COPY_CHUNK_CASE(ColumnVector<Path>)
         COPY_CHUNK_CASE(ColumnVector<EntityList>)
 
         default: {
@@ -87,7 +84,6 @@ inline void copyTransformedChunkImpl(const ColumnVector<size_t>* transform,
         COPY_TRANSFORMED_CHUNK_CASE(ColumnOptVector<types::Double::Primitive>)
         COPY_TRANSFORMED_CHUNK_CASE(ColumnOptVector<types::String::Primitive>)
         COPY_TRANSFORMED_CHUNK_CASE(ColumnOptVector<types::Bool::Primitive>)
-        COPY_TRANSFORMED_CHUNK_CASE(ColumnVector<Path>)
         COPY_TRANSFORMED_CHUNK_CASE(ColumnVector<EntityList>)
         default: {
             bioassert(false, "copyTransformedChunk operator not handled between columns of kind {} and {}",
