+++ query/plan/nodes/PathExplorerNode.h
